tests/frontend/requirements.txt
pytest==8.0.0
pytest-playwright==0.4.4
pytest-html==4.1.1
python-dotenv==1.0.1
